Move transposition table to own file
Bench: 4489497
